backend/requirements.txt
fastapi==0.109.0
python-multipart==0.0.6
uvicorn==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.26.3
PyPDF2==3.0.1
python-docx==1.0.1
docx2txt==0.8
python-pptx==0.6.22
pydantic==2.9.2
faiss-cpu==1.7.4
langchain-openai==0.2.3
typing-extensions==4.12.2
PyMuPDF==1.24.13
langchain-community==0.3.5
pymongo==4.10.1
pyodbc==5.1.0
SQLAlchemy==2.0.31
llama-index==0.12.2
llama-index-core==0.12.2
llama-index-embeddings-azure-openai==0.3.0
llama-index-llms-azure-openai==0.3.0
nltk==3.9.1
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
mammoth==1.9.0
 